% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_RGWAS.R
\name{cumsumPos}
\alias{cumsumPos}
\title{Function to calculate cumulative position (beyond chromosome)}
\usage{
cumsumPos(map)
}
\arguments{
\item{map}{Data frame with the marker names in the first column. The second and third columns contain the chromosome and map position.}
}
\value{
Cumulative position (beyond chromosome) will be returned.
}
\description{
Function to calculate cumulative position (beyond chromosome)
}
