% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_RGWAS.R
\name{design.Z}
\alias{design.Z}
\title{Function to generate design matrix (Z)}
\usage{
design.Z(pheno.labels, geno.names)
}
\arguments{
\item{pheno.labels}{A vector of genotype (line; accesion; variety) names which correpond to phenotypic values.}

\item{geno.names}{A vector of genotype (line; accesion; variety) names for marker genotype data (duplication is not recommended).}
}
\value{
Z of \eqn{y = X\beta + Zu + e}. Design matrix, which is useful for GS or GWAS.
}
\description{
Function to generate design matrix (Z)
}
