% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_RGWAS.R
\name{genesetmap}
\alias{genesetmap}
\title{Function to generate map for gene set}
\usage{
genesetmap(map, gene.set, cumulative = FALSE)
}
\arguments{
\item{map}{Data frame with the marker names in the first column. The second and third columns contain the chromosome and map position.}

\item{gene.set}{Gene information with the format of a "data.frame" (whose dimension is (the number of gene) x 2).
In the first column, you should assign the gene name. And in the second column, you should assign the names of each marker,
which correspond to the marker names of "map" argument.}

\item{cumulative}{If this argument is TRUE, cumulative position will be returned.}
}
\value{
Map for gene set.
}
\description{
Function to generate map for gene set
}
