% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_RGWAS.R
\name{is.diag}
\alias{is.diag}
\title{Function to judge the square matrix whether it is diagonal matrix or not}
\usage{
is.diag(x)
}
\arguments{
\item{x}{Square matrix.}
}
\value{
If `x` is diagonal matrix, `TRUE`. Otherwise the function returns `FALSE`.
}
\description{
Function to judge the square matrix whether it is diagonal matrix or not
}
