% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_RGWAS.R
\name{make.full}
\alias{make.full}
\title{Change a matrix to full-rank matrix}
\usage{
make.full(X)
}
\arguments{
\item{X}{A \eqn{n \times p} matrix which you want to change into full-rank matrix.}
}
\value{
A full-rank matrix
}
\description{
Change a matrix to full-rank matrix
}
