% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{make_hextile}
\alias{make_hextile}
\title{Creates a Tiling of LGBT Hex Stickers}
\usage{
make_hextile(data, width = 1000)
}
\arguments{
\item{data}{data frame or tibble specifying the tiling}

\item{width}{width of each hex tile in pixels}
}
\description{
Creates a Tiling of LGBT Hex Stickers
}
