% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_RGWAS.R
\name{manhattan.plus}
\alias{manhattan.plus}
\title{Add points of -log10(p) corrected by kernel methods to manhattan plot}
\usage{
manhattan.plus(
  input,
  checks,
  cex = 1,
  plot.col1 = c("dark blue", "cornflowerblue"),
  plot.col3 = c("red3", "orange3"),
  plot.type = "p",
  plot.pch = 16
)
}
\arguments{
\item{input}{Data frame of GWAS results where the first column is the marker names,
the second and third column is the chromosome amd map position, and the forth column is -log10(p) for each marker.}

\item{checks}{The marker numbers whose -log10(p)s are corrected by kernel methods.}

\item{cex}{A numerical value giving the amount by which plotting text and symbols should be magnified relative to the default.}

\item{plot.col1}{This argument determines the color of the manhattan plot.
You should substitute this argument as a color vector whose length is 2.
plot.col1[1] for odd chromosomes and plot.col1[2] for even chromosomes.}

\item{plot.col3}{Color of -log10(p) corrected by kernel methods. plot.col3[1] for odd chromosomes and plot.col3[2] for even chromosomes}

\item{plot.type}{This argument determines the type of the manhattan plot. See the help page of "plot".}

\item{plot.pch}{This argument determines the shape of the dot of the manhattan plot. See the help page of "plot".}
}
\value{
Draw manhttan plot
}
\description{
Add points of -log10(p) corrected by kernel methods to manhattan plot
}
