% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_RGWAS.R
\name{manhattan2}
\alias{manhattan2}
\title{Draw manhattan plot (another method)}
\usage{
manhattan2(
  input,
  sig.level = 0.05,
  method.thres = "BH",
  cex = 1,
  plot.col2 = 1,
  plot.type = "p",
  plot.pch = 16,
  cum.pos = NULL,
  lwd.thres = 1,
  cex.lab = 1,
  cex.axis = 1
)
}
\arguments{
\item{input}{Data frame of GWAS results where the first column is the marker names,
the second and third column is the chromosome amd map position, and the forth column is -log10(p) for each marker.}

\item{sig.level}{Siginifincance level for the threshold. The default is 0.05.}

\item{method.thres}{Method for detemining threshold of significance. "BH" and "Bonferroni are offered.}

\item{cex}{A numerical value giving the amount by which plotting text and symbols should be magnified relative to the default.}

\item{plot.col2}{Color of the manhattan plot. color changes with chromosome and it starts from plot.col2 + 1
(so plot.col2 = 1 means color starts from red.)}

\item{plot.type}{This argument determines the type of the manhattan plot. See the help page of "plot".}

\item{plot.pch}{This argument determines the shape of the dot of the manhattan plot. See the help page of "plot".}

\item{cum.pos}{Cumulative position (over chromosomes) of each marker}

\item{lwd.thres}{The line width for the threshold.}

\item{cex.lab}{The font size of the labels.}

\item{cex.axis}{The font size of the axes.}
}
\value{
Draw manhttan plot
}
\description{
Draw manhattan plot (another method)
}
