% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGWAS.menu.R
\name{RGWAS.menu}
\alias{RGWAS.menu}
\title{Print the R code which you should perform for RAINBOWR GWAS}
\usage{
RGWAS.menu()
}
\value{
The R code which you should perform for RAINBOWR GWAS
}
\description{
Print the R code which you should perform for RAINBOWR (Reliable Association INference By Optimizing Weights with R).
}
