% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rice_geno_map.R
\name{Rice_geno_map}
\alias{Rice_geno_map}
\title{Physical map of rice genome}
\format{
A data frame with 1311 rows and 3 variables:
\describe{
  \item{marker}{marker name for each marker, character}
  \item{chr}{chromosome number for each marker, integer}
  \item{pos}{physical position for each marker, integer, (b.p.)}
}
}
\source{
\url{http://www.ricediversity.org/data/}
}
\description{
A dataset containing the information of phycical map of rice genome (Zhao et al., 2010; PLoS One 5(5): e10780).
}
\references{
Zhao K, Wright M, Kimball J, Eizenga G, McClung A, Kovach M, Tyagi W, Ali ML, Tung CW, Reynolds A, Bustamante CD, McCouch SR (2010). Genomic Diversity and Introgression in O. sativa Reveal the Impact of Domestication and Breeding on the Rice Genome. PLoS One. 2010; 5(5): e10780.
}
