% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rice_geno_score.R
\name{Rice_geno_score}
\alias{Rice_geno_score}
\title{Marker genotype of rice genome}
\format{
A data frame with 1311 rows and 395 variables:
\describe{
  Each column shows the marker genotype of each accession.
  The column names are the names of accessions and the rownames are the names of markers.
}
}
\source{
\url{http://www.ricediversity.org/data/}
}
\description{
A dataset containing the information of marker genotype (scored with {-1, 0, 1})
 of rice genome (Zhao et al., 2010; PLoS One 5(5): e10780).
}
\references{
Zhao K, Wright M, Kimball J, Eizenga G, McClung A, Kovach M, Tyagi W, Ali ML, Tung CW, Reynolds A, Bustamante CD, McCouch SR (2010). Genomic Diversity and Introgression in O. sativa Reveal the Impact of Domestication and Breeding on the Rice Genome. PLoS One. 2010; 5(5): e10780.
}
