% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rice_pheno.R
\name{Rice_pheno}
\alias{Rice_pheno}
\title{Phenotype data of rice field trial}
\format{
A data frame with 413 rows and 36 variables:
\describe{
  Phenotypic data of 36 traits obtained by the field trial with 413 genotypes.
}
}
\source{
\url{http://www.ricediversity.org/data/}
}
\description{
A dataset containing the information of phenotype data of rice field trial (Zhao et al., 2011; Nat Comm 2:467).
}
\references{
Zhao, K. et al. (2011) Genome-wide association mapping reveals a rich genetic architecture of complex traits in Oryza sativa. Nat Commun. 2: 467.
}
