% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_useful_functions.R
\name{See}
\alias{See}
\title{Function to view the first part of data (like head(), tail())}
\usage{
See(
  data,
  fh = TRUE,
  fl = TRUE,
  rown = 6,
  coln = 6,
  rowst = 1,
  colst = 1,
  narray = 2,
  drop = FALSE,
  save.variable = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{Your data. 'vector', 'matrix', 'array' (whose dimensions <= 4), 'data.frame' are supported format.
If other formatted data is assigned, str(data) will be returned.}

\item{fh}{From head. If this argument is TRUE, first part (row) of data will be shown (like head() function).
If FALSE, last part (row) of your data will be shown (like tail() function).}

\item{fl}{From left. If this argument is TRUE, first part (column) of data will be shown (like head() function).
If FALSE, last part (column) of your data will be shown (like tail() function).}

\item{rown}{The number of rows shown in console.}

\item{coln}{The number of columns shown in console.}

\item{rowst}{The start point for the direction of row.}

\item{colst}{The start point for the direction of column.}

\item{narray}{The number of dimensions other than row and column shown in console.
This argument is effective only your data is array (whose dimensions >= 3).}

\item{drop}{When rown = 1 or coln = 1, the dimension will be reduced if this argument is TRUE.}

\item{save.variable}{If you want to assign the result to a variable, please set this agument TRUE.}

\item{verbose}{If TRUE, print the first part of data.}
}
\value{
If save.variable is FALSE, NULL. If TRUE, the first part of your data will be returned.
}
\description{
Function to view the first part of data (like head(), tail())
}
