% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_useful_functions.R
\name{estPhylo}
\alias{estPhylo}
\title{Function to estimate & plot phylogenetic tree}
\usage{
estPhylo(blockInterest = NULL, gwasRes = NULL, nTopRes = 1,
  gene.set = NULL, indexRegion = 1:10, chrInterest = NULL,
  posRegion = NULL, pheno, geno, ZETA = NULL, plotTree = TRUE,
  distMat = NULL, distMethod = "manhattan", evolutionDist = FALSE,
  subpopInfo = NULL, groupingMethod = "kmeans", nGrp = 4,
  nIterClustering = 100, saveName = NULL, saveStyle = "png",
  pchBase = c(1, 16), colNodeBase = c(2, 4), colTipCand = 2:8,
  cexMax = 2, edgeColoring = TRUE, tipLabel = TRUE)
}
\arguments{
\item{tipLabel}{}

\item{groupingMethod@param}{}
}
\value{
\describe{
\item{$gvTotal}{a}
\item{$minuslog10p}{a}
}
}
\description{
Function to estimate & plot phylogenetic tree
}
