% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_useful_functions.R
\name{estPhylo}
\alias{estPhylo}
\title{Function to estimate & plot phylogenetic tree}
\usage{
estPhylo(blockInterest = NULL, gwasRes = NULL, nTopRes = 1,
  gene.set = NULL, indexRegion = 1:10, chrInterest = NULL,
  posRegion = NULL, blockName = NULL, pheno = NULL, geno = NULL,
  ZETA = NULL, plotTree = TRUE, distMat = NULL,
  distMethod = "manhattan", evolutionDist = FALSE, subpopInfo = NULL,
  groupingMethod = "kmedoids", nGrp = 4, nIterClustering = 100,
  kernelType = "A.mat", saveName = NULL, saveStyle = "png",
  pchBase = c(1, 16), colNodeBase = c(2, 4), colTipBase = c(3, 5, 6,
  7), cexMax = 2, edgeColoring = TRUE, tipLabel = TRUE,
  verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\value{
\describe{
\item{$gvTotal}{a}
\item{$minuslog10p}{a}
}
}
\description{
Function to estimate & plot phylogenetic tree
}
