% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{make_banner}
\alias{make_banner}
\title{Creates an LGBT Pride Flag with R Logo}
\usage{
make_banner(flag, width = 1000)
}
\arguments{
\item{flag}{character vector with names of the flags}

\item{width}{numeric (width of the flag)}
}
\description{
Creates an LGBT Pride Flag with R Logo
}
