% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{make_hex}
\alias{make_hex}
\title{Creates an LGBT Pride Hex Sticker with R Logo}
\usage{
make_hex(flag = "rainbow", width = 1000)
}
\arguments{
\item{flag}{character string (e.g., rainbow) naming the flag to be drawn}

\item{width}{numeric (width of the flag in pixels)}
}
\value{
An object of class "magick-image"
}
\description{
Creates an LGBT Pride Hex Sticker with R Logo
}
