% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_for_RGWAS.R
\name{manhattan3}
\alias{manhattan3}
\title{Draw the effects of epistasis (3d plot and 2d plot)}
\usage{
manhattan3(
  input,
  cum.pos,
  plot.epi.3d = TRUE,
  plot.epi.2d = TRUE,
  main.epi.3d = NULL,
  main.epi.2d = NULL,
  saveName = NULL
)
}
\arguments{
\item{input}{Data frame of GWAS results where the first column is the marker names,
the second and third column is the chromosome amd map position, and the forth column is -log10(p) for each marker.}

\item{cum.pos}{Cumulative position (over chromosomes) of each marker}

\item{plot.epi.3d}{If TRUE, draw 3d plot}

\item{plot.epi.2d}{If TRUE, draw 2d plot}

\item{main.epi.3d}{The title of 3d plot. If this argument is NULL, trait name is set as the title.}

\item{main.epi.2d}{The title of 2d plot. If this argument is NULL, trait name is set as the title.}

\item{saveName}{When drawing any plot, you can save plots in png format. In saveName, you should substitute the name you want to save.
When saveAt = NULL, the plot is not saved.}
}
\value{
Draw 3d plot and 2d plot to show epistatic effects
}
\description{
Draw the effects of epistasis (3d plot and 2d plot)
}
