% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAINBOWR.R
\docType{package}
\name{RAINBOWR}
\alias{RAINBOWR}
\title{RAINBOWR:
Perform Genome-Wide Asscoiation Study (GWAS) By Kernel-Based Methods}
\description{
By using 'RAINBOWR' (Reliable Association INference By Optimizing Weights with R), users can test multiple SNPs (Single Nucleotide Polymorphisms) simultaneously by kernel-based (SNP-set) methods.
Users can test not only additive effects but also dominance and epistatic effects.
In detail, please check our preprint on bioRxiv: Kosuke Hamazaki and Hiroyoshi Iwata (2019) <doi:10.1101/612028>.
}
