% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rice_Zhao_etal.R
\docType{data}
\name{Rice_Zhao_etal}
\alias{Rice_Zhao_etal}
\title{Rice_Zhao_etal:}
\format{
A list of 3 data frames:
\describe{
  \item{$genoScore}{marker genotyope, \link{Rice_geno_score}}
  \item{$genoMap}{physical map, \link{Rice_geno_map}}
  \item{$pheno}{phenotype, \link{Rice_pheno}}
}
}
\source{
\url{http://www.ricediversity.org/data/}
}
\usage{
Rice_Zhao_etal
}
\description{
A list containing the information of marker genotype of rice genome (Zhao et al., 2010; PLoS One 5(5): e10780)
and phenotype data of rice field trial (Zhao et al., 2011; Nat Comm 2:467).
}
\details{
Marker genotype and phenotype data of rice by Zhao et al., 2010.
}
\references{
Zhao K, Wright M, Kimball J, Eizenga G, McClung A, Kovach M, Tyagi W, Ali ML, Tung CW, Reynolds A, Bustamante CD, McCouch SR (2010). Genomic Diversity and Introgression in O. sativa Reveal the Impact of Domestication and Breeding on the Rice Genome. PLoS One. 2010; 5(5): e10780.
Zhao, K. et al. (2011) Genome-wide association mapping reveals a rich genetic architecture of complex traits in Oryza sativa. Nat Commun. 2: 467.
}
\seealso{
\link{Rice_geno_score}, \link{Rice_geno_map}, \link{Rice_pheno}
}
\keyword{datasets}
